do not enable server statistics for agent instances
The arangodb starter seems to set the `--server.statistics false`
parameter for agents, which disables all statistics.
The original motivation for this was to turn off a background thread
in the instance that periodically writes statistics into RocksDB.
This background writing of the statistics history causes load and
more disk space usage, which we want to avoid on agent instances.

However, setting `--server.statistics false` not only disables the
background writing of statistics into RocksDB, but also the statistics
API and metrics for process statistics.

As it was not ideal to have a single startup option to control two
different things, since 3.4.9 arangod provides a separate option
`--server.statistics-history`. This option controls whether the
statistics are written to RocksDB. This is what we want to turn off.

The good thing is that when this option is not set, it automatically
turns itself off on agents and DB servers.

The defaults in arangod are:
* `--server.statistics`: always `true`, for all instance types
* `--server.statistics-history`: `true` for coordinators, single servers
  and active failover instances, `false` for agents and DB servers.

These are already the optimal settings, so that the arangodb starter
does not need to set any statistics-related startup options for arangod.

This PR modifies the starter to not set any of the above arangod startup
options, as the arangod defaults seem good enough since at least 3.4.9.